E2EUI-1216 tests

+++ src/test/java/co/uk/gel/proj/steps/TestHooks.java
@@ -9,7 +9,10 @@ import io.cucumber.core.event.Status;
 import io.cucumber.java.Before;
 import io.cucumber.java.After;
 import org.openqa.selenium.*;
-import org.openqa.selenium.support.FindBy;
+
+import static co.uk.gel.lib.Actions.acceptAlert;
+import static co.uk.gel.lib.Actions.isAlertPresent;
+
 
 public class TestHooks extends Pages {
 
@@ -85,6 +88,9 @@ public class TestHooks extends Pages {
         Debugger.println("deleted cookies");
         driver.findElement(By.xpath("//a[text()='Log out']")).click(); // Logging out to restart new session
         try {
+           if(isAlertPresent(driver)){
+                acceptAlert(driver);
+            }
             driver.manage().deleteAllCookies();
         } catch (UnhandledAlertException f) {
             try {
